[ADD] Added write and cancel methods for mandate.

+++ V12/twikey_integration/wizard/wizard_cancel_reason.py
@@ -22,20 +22,23 @@ class MandateCancelReason(models.TransientModel):
     
     def action_cancel_confirm(self):
         context = self.env.context
-        host_url = 'https://api.beta.twikey.com/creditor/mandate'
+        base_url=self.env['ir.config_parameter'].sudo().get_param(
+                'twikey_integration.base_url')
         authorization_token=self.env['ir.config_parameter'].sudo().get_param(
                 'twikey_integration.authorization_token')
         if authorization_token:
-            prepared_url = host_url + '?mndtId=' + self.mandate_id.reference + '&rsn=' + self.name
-            response = requests.delete(prepared_url, headers={'Authorization' : authorization_token})
-            if response.status_code == 200:
-                self.mandate_id.update_feed()
-                if self.mandate_id.state == 'signed':
-                    self.mandate_id.write({'state' : 'cancelled', 'description' : self.name})
-                if self.mandate_id.state == 'pending':
-                    self.mandate_id.unlink()
-                    return self.env.ref('twikey_integration.mandate_details_action').read()[0]
-            else:
-                resp_obj = response.json()
-                raise UserError(_('%s')
-                            % (resp_obj.get('message')))
+            prepared_url = base_url + '/creditor/mandate' + '?mndtId=' + self.mandate_id.reference + '&rsn=' + self.name
+            self.mandate_id.update_feed()
+            if self.mandate_id.state != 'cancelled':
+                response = requests.delete(prepared_url, headers={'Authorization' : authorization_token})
+                if response.status_code == 200:
+    #                 self.mandate_id.update_feed()
+                    if self.mandate_id.state == 'signed':
+                        self.mandate_id.write({'state' : 'cancelled', 'description' : self.name})
+                    if self.mandate_id.state == 'pending':
+                        self.mandate_id.unlink()
+                        return self.env.ref('twikey_integration.mandate_details_action').read()[0]
+                else:
+                    resp_obj = response.json()
+                    raise UserError(_('%s')
+                                % (resp_obj.get('message')))
